src: rename 'ansible' directory
Fixes: https://github.com/oVirt/ovirt-hosted-engine-setup/issues/45
