+++ src/he_ansible/callback_plugins/1_otopi_json.py
