+++ src/he_ansible/callback_plugins/2_ovirt_logger.py
